Streamline some style properties
