+++ packages/alfa-css/src/value/shadow.ts
@@ -4,6 +4,11 @@ import { Length } from "./numeric";
 
 import { Value } from "../value";
 import { Color } from "./color";
+import { Parser } from "@siteimprove/alfa-parser";
+import { Slice } from "@siteimprove/alfa-slice";
+import { Token } from "../syntax";
+import { Keyword } from "./keyword";
+import { Err, Result } from "@siteimprove/alfa-result";
 
 /**
  * @public
@@ -136,4 +141,106 @@ export namespace Shadow {
     color: Color.JSON;
     isInset: boolean;
   }
+
+  export const parse: Parser<Slice<Token>, Shadow, string> = (input) => {
+    let horizontal: Length.Fixed | undefined;
+    let vertical: Length.Fixed | undefined;
+    let blur: Length.Fixed | undefined;
+    let spread: Length.Fixed | undefined;
+    let color: Color | undefined;
+    let isInset: boolean | undefined;
+
+    const skipWhitespace = () => {
+      for (const [remainder] of Token.parseWhitespace(input)) {
+        input = remainder;
+      }
+    };
+
+    while (true) {
+      skipWhitespace();
+
+      if (horizontal === undefined) {
+        // horizontal: <length>
+        const result = Length.parseBase(input);
+
+        if (result.isOk()) {
+          [input, horizontal] = result.get();
+          skipWhitespace();
+
+          {
+            // vertical: <length>
+            const result = Length.parseBase(input);
+
+            if (result.isErr()) {
+              return result;
+            }
+
+            // the previous check ensure that the result is Ok
+            [input, vertical] = result.getUnsafe();
+            skipWhitespace();
+
+            {
+              // blur: <length>?
+              const result = Length.parseBase(input);
+
+              if (result.isOk()) {
+                [input, blur] = result.get();
+                skipWhitespace();
+
+                {
+                  // spread: <length>?
+                  const result = Length.parseBase(input);
+
+                  if (result.isOk()) {
+                    [input, spread] = result.get();
+                  }
+                }
+              }
+            }
+          }
+
+          continue;
+        }
+      }
+
+      if (color === undefined) {
+        // color: <color>?
+        const result = Color.parse(input);
+
+        if (result.isOk()) {
+          [input, color] = result.get();
+          continue;
+        }
+      }
+
+      if (isInset === undefined) {
+        // isInset: inset?
+        const result = Keyword.parse("inset")(input);
+
+        if (result.isOk()) {
+          isInset = true;
+          [input] = result.get();
+          continue;
+        }
+      }
+
+      break;
+    }
+
+    if (horizontal === undefined || vertical === undefined) {
+      return Err.of("Expected horizontal and vertical offset");
+    }
+
+    return Result.of([
+      input,
+      Shadow.of(
+        horizontal,
+        vertical,
+        blur ?? Length.of(0, "px"),
+        spread ?? Length.of(0, "px"),
+        color ?? Keyword.of("currentcolor"),
+        isInset ?? false
+      ),
+    ]);
+  };
 }
