+++ packages/alfa-style/src/property/border-top-width.ts
@@ -1,5 +1,4 @@
-import { Keyword } from "@siteimprove/alfa-css";
-import { Length } from "@siteimprove/alfa-css/src/value/numeric";
+import { Keyword, Length } from "@siteimprove/alfa-css";
 import { Parser } from "@siteimprove/alfa-parser";
 
 import { Longhand } from "../longhand";
